Basic arena spawning
On BeginPlay of the Game Mode we activate the first Arena that is found in the level. This triggers a function in the arena that reads from out spawn DT and loads all soft object references for each spanable area. Currently there is a test level with 1 arena, two spawnable areas and there is 1 entry in the spawn DT.